Trabajo 10 Dic

+++ firmware_v3/libs/touch/inc/XPT2046_Touchscreen.h
@@ -46,6 +46,6 @@ bool_t XPT2046_Touchscreen_begin(XPT2046_Touchscreen *touchscreen);
 TS_Point XPT2046_Touchscreen_getPoint(XPT2046_Touchscreen *touchscreen);
 bool_t XPT2046_Touchscreen_touched(XPT2046_Touchscreen *touchscreen);
 void XPT2046_Touchscreen_update(XPT2046_Touchscreen *touchscreen);
-void XPT2046_Touchscreen_readData(XPT2046_Touchscreen *touchscreen, uint16_t *x, uint16_t *y, uint8_t *z);
+void XPT2046_Touchscreen_readData(XPT2046_Touchscreen *touchscreen, uint16_t *x, uint16_t *y, uint16_t *z);
 
 #endif // XPT2046_TOUCHSCREEN_H
